+++ src/main/java/org/truffleruby/core/CoreLibrary.java
@@ -946,7 +946,6 @@ public class CoreLibrary {
             "/core/pre.rb",
             "/core/basic_object.rb",
             "/core/array.rb",
-            "/core/channel.rb",
             "/core/configuration.rb",
             "/core/false.rb",
             "/core/fiber.rb",
@@ -1053,6 +1052,7 @@ public class CoreLibrary {
             "/core/posix.rb",
             "/core/main.rb",
             "/core/post.rb",
+            "/core/truffle/queue_operations.rb",
             POST_BOOT_FILE
     };