+++ src/main/java/org/truffleruby/core/thread/ThreadManager.java
@@ -564,14 +564,14 @@ public class ThreadManager {
         }
     }
 
-    /** Runs {@code action} until it returns a non-null value. The given action should throw an
-     * {@link InterruptedException} when {@link Thread#interrupt()} is called. Otherwise, the {@link SafepointManager}
-     * will not be able to interrupt this action. See {@link ThreadNodes.ThreadRunBlockingSystemCallNode} for blocking
-     * native calls. If the action throws an {@link InterruptedException}, it will be retried until it returns a
-     * non-null value.
+    /** Runs {@code action} until it returns a value (IOW, until it does not throw {@link InterruptedException}). The
+     * given action should throw an {@link InterruptedException} when {@link Thread#interrupt()} is called. Otherwise,
+     * the {@link SafepointManager} will not be able to interrupt this action. See
+     * {@link ThreadNodes.ThreadRunBlockingSystemCallNode} for blocking native calls. If the action throws an
+     * {@link InterruptedException}, it will be retried until it returns a non-null value.
      *
      * @param action must not touch any Ruby state
-     * @return the first non-null return value from {@code action} */
+     * @return the return value from {@code action} */
     @TruffleBoundary
     public <T> T runUntilResult(Node currentNode, BlockingAction<T> action) {
         return runUntilResult(currentNode, action, null, null);